requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
attrs==25.4.0
blockbuster==1.5.25
bracex==2.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cryptography==44.0.3
daytona==0.115.2
daytona_api_client==0.115.2
daytona_api_client_async==0.115.2
daytona_toolbox_api_client==0.115.2
daytona_toolbox_api_client_async==0.115.2
deepagents==0.2.7
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
environs==14.5.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
langchain==1.0.7
langchain-anthropic==1.1.0
langchain-core==1.0.5
langchain-mcp-adapters==0.1.13
langchain-openai==1.0.3
langgraph==1.0.3
langgraph-api==0.5.20
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.4
langgraph-runtime-inmem==0.18.0
langgraph-sdk==0.2.9
langsmith==0.4.43
markdown-it-py==4.0.0
marshmallow==4.1.0
mcp==1.21.2
mdurl==0.1.2
multidict==6.7.0
multipart==1.3.0
obstore==0.7.3
openai==2.8.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
propcache==0.4.1
protobuf==6.33.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.29.0
runloop_api_client==0.68.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.1.3
starlette==0.50.0
structlog==25.5.0
tavily==1.1.0
tavily-python==0.7.13
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.1
truststore==0.10.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uuid_utils==0.11.1
uvicorn==0.38.0
watchfiles==1.1.1
wcmatch==10.1
websockets==15.0.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
